+++ Module.h
@@ -40,7 +40,7 @@ public:
 
     void ActivateSymbol(const std::string& word, int16_t K);
 
-    void AddExcitationToIndex(uint32_t index, float value);
+    void AddExcitationToIndex(uint16_t index, float value);
     void AddExcitationToAllSymbols(int16_t K);
 
     void AddExcitationVector(const IExcitationVector<float>& input);
@@ -65,9 +65,9 @@ private:
     std::unique_ptr<IExcitationVector<float>> excitations_;
     std::unique_ptr<IExcitationVector<float>> normalized_excitations_;
     std::unique_ptr<IExcitationVector<uint32_t>> kb_inputs_;
-    std::unique_ptr<std::set<uint32_t>> frozen_indexes_;
+    std::unique_ptr<std::set<uint16_t>> frozen_indexes_;
 
-    std::unique_ptr<std::pair<uint32_t, float>> MaxExcitation(const std::set<std::pair<uint16_t, float> > &nz_excitations);
+    std::unique_ptr<std::pair<uint16_t, float>> MaxExcitation(const std::set<std::pair<uint16_t, float> > &nz_excitations);
     std::set<std::pair<uint16_t, float>> ExcitationsAbove(int8_t K, const std::set<std::pair<uint16_t, float>>& nz_excitations);
     int16_t ActualK(int16_t K);
     int16_t MaxK();
